refactor(frontend): move editor screen into feature folder
